+++ internal/provider/data_source_fmc_ftd_nat_policy_test.go
@@ -33,7 +33,7 @@ func TestAccDataSourceFmcFTDNATPolicy(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_ftd_nat_policy.test", "name", "my_ftd_nat_policy"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_ftd_nat_policy.test", "description", "My nat policy"))
 	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_ftd_nat_policy.test", "type"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_ftd_nat_policy.test", "manual_nat_rules.0.description", "My manual nat rule 1"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_ftd_nat_policy.test", "manual_nat_rules.0.description", "My Manual NAT rule 1"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_ftd_nat_policy.test", "manual_nat_rules.0.enabled", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_ftd_nat_policy.test", "manual_nat_rules.0.section", "BEFORE_AUTO"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_ftd_nat_policy.test", "manual_nat_rules.0.nat_type", "STATIC"))
@@ -89,7 +89,7 @@ func testAccDataSourceFmcFTDNATPolicyConfig() string {
 	config += `	description = "My nat policy"` + "\n"
 	config += `	manage_rules = true` + "\n"
 	config += `	manual_nat_rules = [{` + "\n"
-	config += `		description = "My manual nat rule 1"` + "\n"
+	config += `		description = "My Manual NAT rule 1"` + "\n"
 	config += `		enabled = true` + "\n"
 	config += `		section = "BEFORE_AUTO"` + "\n"
 	config += `		nat_type = "STATIC"` + "\n"
@@ -119,7 +119,7 @@ func testAccNamedDataSourceFmcFTDNATPolicyConfig() string {
 	config += `	description = "My nat policy"` + "\n"
 	config += `	manage_rules = true` + "\n"
 	config += `	manual_nat_rules = [{` + "\n"
-	config += `		description = "My manual nat rule 1"` + "\n"
+	config += `		description = "My Manual NAT rule 1"` + "\n"
 	config += `		enabled = true` + "\n"
 	config += `		section = "BEFORE_AUTO"` + "\n"
 	config += `		nat_type = "STATIC"` + "\n"